pinataSignedUrl should return the date and expires fields as well
